backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
requests
groq
python-dotenv
